Fix threading issues: protect shared state with Controller::lock_
- Added MMThreadGuard to all property action handlers
- Protected ReadDeviceStateByMode, ReadCurrentDeviceState, etc.
- Fixed potential data race between polling thread and core thread

+++ DeviceAdapters/3Z_Optics/3Z_Optics.h
@@ -116,7 +116,7 @@ private:
 	int ReadAllChannelRegisters();
 	int ReadDeviceStateByMode(int mode);
 	void UpdatePropertiesFromDevice();
-	uint16_t CalculateCRC(const uint8_t* data, int length);
+	uint16_t CalculateCRC(const uint8_t* data, size_t length);
 	int SendModbusCommand(const std::vector<uint8_t>& request, std::vector<uint8_t>& response, int expectedResponseLength);
 };
